Create new api UI for web socket api
