+++ features/apimgt/org.wso2.carbon.apimgt.publisher.feature/src/main/resources/publisher/public/locales/en.json
@@ -3,6 +3,7 @@
   "wso2.api.publisher.enables.api.providers.to.publish.apis,.share.documentation,.provision.api.keys.and.gather.feedback.on.features,.quality.and.usage..to.get.started,.create.an.api.or.publish.a.sample.api..": "WSO2 API Publisher enables API providers to publish APIs, share documentation, provision API keys and gather feedback on features, quality and usage. To get started, Create an API or Publish a sample API.",
   "create.new.api": "Create New API",
   "create.new.rest.api": "Create New REST API",
+  "create.new.websocket.api": "Create New WebSocket API",
   "create.new.api.swagger": "Create New API - ",
   "swagger.file.upload": "Swagger file upload",
   "by.swagger.url": "By swagger url",
@@ -14,14 +15,17 @@
   "i.have.an.existing.api": "I Have an Existing API",
   "use.an.existing.api.endpoint.or.the.api.swagger.definition.to.create.an.api": "Use an existing API endpoint or the API Swagger definition to create an API",
   "i.have.a.soap.endpoint": "I Have a SOAP Endpoint",
+  "use.an.existing.soap.or.import.the.wsdl": "Use an existing SOAP or Import the WSDL",
   "use.an.existing.soap.endpoint.to.create.a.managed.api.import.the.wsdl.of.the.soap.service": "Use an existing SOAP endpoint to create a managed API. Import the WSDL of the SOAP service",
   "design.a.new.rest.api": "Design a New REST API",
+  "i.have.an.existing.api": "I Have an Existing REST API",
+  "use.an.existing.rest.endpoint.or.Swagger.definition": "Use an existing REST endpoint or Swagger definition",
   "design.a.new.rest.api.using.wsdl": "Design a new REST API using WSDL",
   "select.wsdl": "Select WSDL",
   "design.and.prototype.a.new.rest.api": "Design and prototype a new REST API",
   "fill.the.mandatory.fields": "Fill the mandatory fields (Name, Version, Context) and create the API. Configure advanced configurations later.",
-  "design.new.websocket.api": "Design New Websocket API",
-  "design.and.prototype.a.new.websocket.api": "Design and prototype a new Websocket API",
+  "design.new.websocket.api": "Design New WebSocket API",
+  "design.and.prototype.a.new.websocket.api": "Design and prototype a new WebSocket API",
   "endpoint": "Endpoint",
   "production": "Production",
   "sandbox": "Sandbox",
